p2p/discover: Relocate constants and helpers to utils.go
