+++ networks/p2p/discover/table_util.go
@@ -18,88 +18,4 @@
 
 package discover
 
-import (
-	crand "crypto/rand"
-	"encoding/binary"
-	"math/rand"
-	"sync"
-	"time"
-)
-
-// TODO: salvage pushNode and deleteNode to this file.
-
-func hasNode(list []*Node, n *Node) bool {
-	for _, e := range list {
-		if e.ID == n.ID {
-			return true
-		}
-	}
-	return false
-}
-
-func bumpNode(list []*Node, n *Node) bool {
-	for i := range list {
-		if list[i].ID == n.ID {
-			// move it to the front
-			copy(list[1:], list[:i])
-			list[0] = n
-			return true
-		}
-	}
-	return false
-}
-
-// A shared pseudorandom source that can be shared by multiple storages.
-// This RNG is shared so that the main table loop can seed it periodically.
-type sharedRand struct {
-	mu   sync.Mutex
-	rand *rand.Rand
-}
-
-func newSharedRand() *sharedRand {
-	return &sharedRand{
-		rand: rand.New(rand.NewSource(time.Now().UnixNano())),
-	}
-}
-
-func (s *sharedRand) Seed() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	var b [8]byte
-	if _, err := crand.Read(b[:]); err != nil {
-		// If crypto/rand seed fails, fall back to time seed
-		s.rand.Seed(time.Now().UnixNano())
-		return
-	}
-	s.rand.Seed(int64(binary.BigEndian.Uint64(b[:])))
-}
-
-func (s *sharedRand) Intn(n int) int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.rand.Intn(n)
-}
-
-func (s *sharedRand) Int63n(n int64) int64 {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.rand.Int63n(n)
-}
-
-func (s *sharedRand) Perm(n int) []int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.rand.Perm(n)
-}
-
-func (s *sharedRand) Shuffle(n int, swap func(i, j int)) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.rand.Shuffle(n, swap)
-}
-
-func (s *sharedRand) Read(p []byte) (n int, err error) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.rand.Read(p)
-}
+// All symbols formerly in this file have been moved to utils.go.
